fix:wrong definition of security flag and move github functions to githubutils repo
